+++ client/store/index.js
@@ -4,11 +4,12 @@ import thunkMiddleware from 'redux-thunk'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import allProducts from './allProducts'
 import singleProduct from './singleProduct'
+import reviews from './productReviews'
 import categories from './categories'
 import currentUser from './currentUser'
 import user from './user'
 
-const reducer = combineReducers({ allProducts, singleProduct, categories, currentUser })
+const reducer = combineReducers({ allProducts, singleProduct, categories, currentUser, reviews })
 const middleware = composeWithDevTools(applyMiddleware(
   thunkMiddleware,
   createLogger({ collapsed: true })
